+++ js/library.js
@@ -42,18 +42,20 @@ function getCurrentChapter()//returns the current chapter as an object
 
 function chapterSize(chapter)//returns the size of a chapter (HAS TO BE MANUALLY CORRECTED)
 {
-    if(chapter == 0)
-    {
-        return 88;
-    }
-    else if (chapter == 1)
-    {
-        return 230;
-    }
-    else if (chapter == 2)
-    {
-        return 300;
-    }
+    return story[0].length;
+
+    // if(chapter == 0)
+    // {
+    //     return 103;
+    // }
+    // else if (chapter == 1)
+    // {
+    //     return 230;
+    // }
+    // else if (chapter == 2)
+    // {
+    //     return 300;
+    // }
 }
 
 //*****CHARACTERS*****